requirements.txt
flask
flask_sqlalchemy
flask_bcrypt
pyjwt